requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
crewai==0.1.30
pydantic==2.6.1
python-multipart==0.0.6